+++ packages/db/prod/src/environment.ts
@@ -1,0 +1,9 @@
+export class MissingDatabaseConnectionError extends Error {
+  constructor() {
+    super(
+      'The environment has not been configured to use Netlify DB. To use it manually, supply the `connectionString` option when calling `getDatabase()`.',
+    )
+
+    this.name = 'MissingDatabaseConnectionError'
+  }
+}