+++ packages/edge-functions/dev/CHANGELOG.md
@@ -1,5 +1,79 @@
 # Changelog
 
+## [1.0.10](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.9...edge-functions-dev-v1.0.10) (2026-02-06)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/runtime-utils bumped from 2.2.1 to 2.3.0
+
+## [1.0.9](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.8...edge-functions-dev-v1.0.9) (2026-01-28)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#576](https://github.com/netlify/primitives/issues/576)) ([1c6ff24](https://github.com/netlify/primitives/commit/1c6ff247318b8cdb1961ce76e96fe423a4ec8cd7))
+
+## [1.0.8](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.7...edge-functions-dev-v1.0.8) (2026-01-19)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#566](https://github.com/netlify/primitives/issues/566)) ([184656a](https://github.com/netlify/primitives/commit/184656af871a81bddc69e9133f22bfa945f28ebf))
+
+## [1.0.7](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.6...edge-functions-dev-v1.0.7) (2025-12-08)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#557](https://github.com/netlify/primitives/issues/557)) ([60644bf](https://github.com/netlify/primitives/commit/60644bf6863a0ba41d9e73b0d8d1cfa90205c963))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev-utils bumped from 4.3.2 to 4.3.3
+    * @netlify/edge-functions bumped from 3.0.2 to 3.0.3
+
+## [1.0.6](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.5...edge-functions-dev-v1.0.6) (2025-12-03)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#549](https://github.com/netlify/primitives/issues/549)) ([ec4fb29](https://github.com/netlify/primitives/commit/ec4fb2978af5cf75c12645ade9f771bc94a796e2))
+
+## [1.0.5](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.4...edge-functions-dev-v1.0.5) (2025-11-14)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/edge-bundler to ^14.8.7 ([#541](https://github.com/netlify/primitives/issues/541)) ([6fee541](https://github.com/netlify/primitives/commit/6fee541e05d67920ce500aafd8e7788c303bb7c0))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev-utils bumped from 4.3.1 to 4.3.2
+
+## [1.0.4](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.3...edge-functions-dev-v1.0.4) (2025-11-05)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#515](https://github.com/netlify/primitives/issues/515)) ([67034cc](https://github.com/netlify/primitives/commit/67034cce9f80e35582f328b1990d0f26e7181cff))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev-utils bumped from 4.3.0 to 4.3.1
+    * @netlify/edge-functions bumped from 3.0.1 to 3.0.2
+
 ## [1.0.3](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.2...edge-functions-dev-v1.0.3) (2025-11-04)